added predict to model types
