+++ portalsite/portal/core/model_type/linear_regression.py
@@ -1,78 +1,20 @@
 # region imports
 # standard
 from json import dumps, loads
-from abc import ABC, abstractmethod
+from typing import TYPE_CHECKING
 
-from typing import TYPE_CHECKING, TypeAlias
 # 3rd party
 import numpy as np
 from sklearn.linear_model import LinearRegression
 
 # local
-from .named_id_manager import NamedIdObject
+from .model_type import ModelStorageType, ModelType
 
 # type hints
 if TYPE_CHECKING:
-    from ..models import Model, Measurement
+    from portal.models import Model, Measurement
 
 # endregion
-
-ModelStorageType: TypeAlias = str
-"""Format/type used to store the model data in the database."""
-class ModelType(ABC,NamedIdObject):
-    """Type of a prediction model"""
-
-    @property
-    @abstractmethod
-    def description(self) -> str:
-        """Description of the type, its uses, and  possbily refering to external docs"""
-
-    @abstractmethod
-    def score(self, model: 'Model', measurement: 'Measurement') -> float:
-        """Returns the score of a measurement against a model"""
-
-    @classmethod
-    @abstractmethod
-    def train(cls,
-              model: 'Model',
-              measurements: list['Measurement'],
-              max_iterations: int,
-              max_seconds: int) -> tuple[ModelStorageType, float]:
-        """Trains a model, returning the new model data with its score"""
-
-class TestModelType(ModelType):
-    """This model type is a dummy implementation for development and testing::
-         - scoring will always return 0
-         - training will always return the same 'model' and 0 score
-    """
-
-    @property
-    def id_(self) -> str:
-        """Identifier used in internal dictionaries - maximal length 10"""
-        return "Test"
-
-    @property
-    def name(self) -> str:
-        """Returns a *short, concise* name"""
-        return "TestModelType"
-
-    @property
-    def description(self) -> str:
-        """Description of the type, its uses, and  possbily refering to external docs"""
-        return self.__doc__
-
-    def score(self, model: 'Model', measurement: 'Measurement') -> float:
-        """Returns the score of a measurement against a model"""
-        return 0.0
-
-    def train(self,
-              model: 'Model',
-              measurements: list['Measurement'],
-              max_iterations: int,
-              max_seconds: int) -> tuple[ModelStorageType, float]:
-        """Trains a model, returning the same model data with a score of 0"""
-        return (model.data, 0.0)
-
 
 class LinearRegressionModel(ModelType):
     """Base implementation of a linear regression model with a dfined nr of features."""
@@ -123,20 +65,21 @@ class LinearRegressionModel(ModelType):
 
     def score(self, model: 'Model', measurement: 'Measurement') -> float:
         """Returns the score of a measurement against a model"""
-        X = measurement.model_input()
-        y = measurement.model_target()
-        # TODO: how do we disintguish labelled and unlabelled measurements? check if y == None?
-        model = self.__load_model(model)
-        print(X)
-        print(y)
-        score = model.score(X, y)
+        # TODO: how do we disintguish labelled and unlabelled measurements? check if model_target == None?
+        model : LinearRegression = self.__load_model(model)
+        score = model.score(measurement.model_input(), measurement.model_target())
         return score
+
+    def predict(self, model: 'Model', measurement: 'Measurement') -> np.ndarray:
+        """Returns the prediction of a measurement against a model"""
+        model : LinearRegression = self.__load_model(model)
+        return model.predict(measurement.model_input())
 
     def train(self,
               model: 'Model',
               measurements: list['Measurement'],
               max_iterations: int,
-              max_seconds: int) -> tuple['Model', float]:
+              max_seconds: int) -> tuple[ModelStorageType, float]:
         """Trains a model, returning the new model with its score"""
         X_list = [m.model_input() for m in measurements]
         y_list = [m.model_target() for m in measurements]
